moving reflections into a wrapper object
