+++ app/src/main/scala/hydra/app/Main.scala
@@ -16,7 +16,7 @@
 
 package hydra.app
 
-import hydra.core.app.BootstrappingSupport
+import hydra.core.bootstrap.BootstrappingSupport
 
 // $COVERAGE-OFF$Disabling highlighting by default until a workaround for https://issues.scala-lang.org/browse/SI-8596 is found
 object Main extends App with BootstrappingSupport {
