+++ core/src/main/scala/hydra/core/bootstrap/ServiceProvider.scala
@@ -1,4 +1,4 @@
-package hydra.core.app
+package hydra.core.bootstrap
 
 import akka.actor.Props
 
@@ -6,7 +6,7 @@ import akka.actor.Props
   * A trait that can be extended by classes providing services (actors) to be loaded during
   * boot time for Hydra.
   *
-  * The [[hydra.core.app.Bootsrapping]] trait scans the classpath for objects that extend this interface
+  * The [[hydra.core.bootstrap.BootstrappingSupport]] trait scans the classpath for objects that extend this interface
   * and adds the services provided to the list of services to be managed by Hydra.
   *
   */
