+++ core/src/test/resources/reference.conf
@@ -3,10 +3,11 @@ application.config.location = ""
 
 constructr {
   coordination {
-    class-name = hydra.core.app.NoOpCoordination
+    class-name = hydra.core.bootstrap.NoOpCoordination
   }
 }
 hydra_test {
+  scan-packages = ["test.scan"]
   schema.registry.url = "mock"
   transports.test-transport.path = /user/test-transport
 }
