+++ core/src/test/scala/hydra/core/bootstrap/NoOpCoordination.scala
@@ -1,4 +1,4 @@
-package hydra.core.app
+package hydra.core.bootstrap
 
 import akka.Done
 import akka.actor.{ActorSystem, Address}
