+++ ingest/src/main/scala/hydra.ingest/services/IngestionActors.scala
@@ -2,7 +2,7 @@ package hydra.ingest.services
 
 import akka.actor.Props
 import hydra.common.util.ActorUtils
-import hydra.core.app.ServiceProvider
+import hydra.core.bootstrap.ServiceProvider
 
 /**
   * Created by alexsilva on 3/29/17.
